Move withProduct to js/hocs
