+++ assets/js/hocs/with-product.js
